An open source API for labelling images and videos
This is an open source API that is able to label images and videos of
various length. It also enables the user to search within the video
using a specific label and the output is the frames where the label has
occurred. The API can be hosted normal PCs and enable the user to
schedule multiple videos to be labelled.